+++ decoder.rb
@@ -1,0 +1,23 @@
+class BaseDecorator < Nameable
+  def initialize(nameable)
+    super
+
+    @nameable = nameable
+  end
+end
+
+class CapitalizeDecorator < BaseDecorator
+  def correct_name
+    @nameable.correct_name.capitalize
+  end
+end
+
+class TrimmerDecorator < BaseDecorator
+  def correct_name
+    if @nameable.correct_name.length > 10
+      @nameable.correct_name[0..9]
+    else
+      @nameable.correct_name
+    end
+  end
+end
