+++ person.rb
@@ -1,14 +1,4 @@
-class Nameable
-  attr_accessor :name
-
-  def initialize(name)
-    @name = name
-  end
-
-  def correct_name
-    raise NotImplementedError, 'Kindly Implement correct_name method to return the correct name'
-  end
-end
+require './namable'
 
 class Person < Nameable
   attr_accessor :name, :age
@@ -37,30 +27,6 @@ class Person < Nameable
 
   def of_age?
     @age >= 18
-  end
-end
-
-class BaseDecorator < Nameable
-  def initialize(nameable)
-    super
-
-    @nameable = nameable
-  end
-end
-
-class CapitalizeDecorator < BaseDecorator
-  def correct_name
-    @nameable.correct_name.capitalize
-  end
-end
-
-class TrimmerDecorator < BaseDecorator
-  def correct_name
-    if @nameable.correct_name.length > 10
-      @nameable.correct_name[0..9]
-    else
-      @nameable.correct_name
-    end
   end
 end
